+++ src/scripts/rule-tests/cases/dcl-ds-inline.ts
@@ -1,0 +1,16 @@
+import type { Case } from '../types';
+
+const testCase: Case = {
+  name: 'dcl-ds-inline',
+  input: [
+    '**free',
+    'dcl-proc Test;',
+    '  dcl-ds ZProcSts likeds(r_ZProcStsDS) inz;',
+    '  end-ds;',
+    '  return;',
+    'end-proc;'
+  ].join('\n'),
+  mustInclude: ['dcl-ds ZProcSts likeds(r_ZProcStsDS) inz; end-ds;']
+};
+
+export default testCase;